+++ lib/$globj.d.ts
@@ -5,8 +5,8 @@
 declare class $globj {
     static readonly document: JQuery;
     static readonly window: JQuery;
-    static readonly html: any;
-    static readonly body: any;
-    static readonly htmlBody: any;
+    static readonly html: JQuery;
+    static readonly body: JQuery;
+    static readonly htmlBody: JQuery;
 }
 export default $globj;