+++ lib/asArray.d.ts
@@ -1,0 +1,6 @@
+/**
+ * Converts a given object/map/set into an array
+ * @param object
+ * @return {Array}
+ */
+export declare function asArray(object: any): Array<any>;