+++ src/forEach.ts
@@ -3,18 +3,21 @@
  * For LABOR.digital
  */
 import {isSet} from "./isSet";
+import {GenericObject} from "./Interfaces";
 
-interface forEachCallbackType extends Function{
+export interface ForEachCallbackType extends Function
+{
 	/**
 	 * Is called for every element of the iterated object
 	 * @param value The current value
 	 * @param key The current key
 	 * @param iteratedObject The iterated object
 	 */
-	(value?, key?:string|number, iteratedObject?):void|boolean
+	(value?, key?: string | number, iteratedObject?): void | boolean
 }
 
-interface forEachCallbackType extends Function{
+export interface ForEachCallbackType extends Function
+{
 	/**
 	 * Is called for every element of the iterated object
 	 * @param $value The current value as a jquery object
@@ -22,11 +25,12 @@ interface forEachCallbackType extends Function{
 	 * @param key The current key
 	 * @param iteratedObject The iterated object
 	 */
-	($value?:JQuery, key?, value?, iteratedObject?):void|boolean
+	($value?: JQuery, key?, value?, iteratedObject?): void | boolean
 }
 
-interface BreakErrorType {
-	breaker?:boolean;
+interface BreakErrorType
+{
+	breaker?: boolean;
 }
 
 const breaker = new Error() as BreakErrorType;
@@ -41,36 +45,43 @@ breaker.breaker = true;
  * @param object The array or object to iterate
  * @param callback The callback to apply. Params are: (value, key, iteratedObject)
  */
-export function forEach(object: Array<any> | { [key: string]: any } | any, callback: forEachCallbackType): void {
-	if (object === null || typeof object === 'undefined') return;
-	if (typeof object.forEach === 'function') {
+export function forEach(object: Array<any> | GenericObject | any, callback: ForEachCallbackType): void
+{
+	if (object === null || typeof object === "undefined") return;
+	if (typeof object.forEach === "function")
+	{
 		const objectIsSet = isSet(object);
-		try {
+		try
+		{
 			let c = 0;
-			object.forEach((v,k) => {
+			object.forEach((v, k) => {
 				if (callback(v, objectIsSet ? c : k, object) === false)
 					throw breaker;
 				++c;
 			});
-		} catch (e) {
-			if(e.breaker === true) return;
+		} catch (e)
+		{
+			if (e.breaker === true) return;
 			throw e;
 		}
 		return;
 	}
-	else if (typeof object.jquery !== 'undefined') {
+	else if (typeof object.jquery !== "undefined")
+	{
+		//noinspection TypeScriptValidateTypes
 		object.each((k, v) => !(callback(jQuery(v), k, v, object) === false));
 		return;
 	}
-	else if(typeof object === 'object') {
-		for (let k in object) {
+	else if (typeof object === "object")
+	{
+		for (let k in object)
+		{
 			if (!object.hasOwnProperty(k)) continue;
-			// @ts-ignore
-			if(parseInt(k)+'' === k) k = parseInt(k);
-			const v = object[k];
-			if (callback(v, k, object) === false) break;
+			let kReal: string|number = k;
+			if (parseInt(k) + "" === k) kReal = parseInt(k);
+			if (callback(object[k], kReal, object) === false) break;
 		}
 		return;
 	}
-	throw Error('Could not iterate given object!');
+	throw Error("Could not iterate given object!");
 }