+++ src/isMap.ts
@@ -1,4 +1,7 @@
-// @todo test if this works with core-js polyfills
+/**
+ * Created by Martin Neundorfer on 20.09.2018.
+ * For LABOR.digital
+ */
 export function isMap(object: any): boolean{
 	if(typeof object === 'undefined' || object === null) return false;
 	return object instanceof Map || object.constructor === 'Map';