+++ CHANGELOG.md
@@ -2,16 +2,25 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+## [1.1.0] - 2018-09-24
 ### Added
 - Added some basic tests
 - Added isMap: To check if a given object is a set or not
 - Added isSet: To check if a given object is a map or not
 - Added mergeRecursive: To merge arrays, objects, sets and Maps with each other.
-- Adds support for map and set on reduce() and map()
+- Added support for map and set on reduce() and map()
+- Added "GenericObject" and "JQueryEvent" interfaces
+- Added "asArray" and "asObject" helpers
+- Added some documentation
+- Added "filter" helper
 
 ### Changed 
 - Converted whole codebase to typescript
 - Renames Globj.js to $globj.js to be consistent
+- Extracts "IterableHelpers" from "forEach"
+
+### Fixed
+- Fixed some issues with include paths when building the library
 
 ## [1.0.3] - 2018-09-06
 ### Added